changed structure of conditions

+++ install.sh
@@ -186,48 +186,57 @@ show "Очистка пакетного менеджера"
   apt autoremove -y && \
   apt autoclean -y
 
+
 # Добавление ключа для авторизации
 echo -en "\n${green}Добавить публичный ключ для авторизации по ssh? [Y/n]: ${end}"
-answer; if [[ $? -ne 0 ]]; then echo ""; fi
-echo -en "\n${cyan}Введите Ваш публичный ключ: ${end}"; read user_pub_key
-mkdir /home/${username}/.ssh
-{
-    echo "${user_pub_key}"
-  } > /home/${username}/.ssh/authorized_keys && \
-  chown ${username}:${username} /home/${username}/.ssh/authorized_keys
+answer
+if [[ $? -eq 0 ]]; then
+  echo -en "\n${cyan}Введите Ваш публичный ключ: ${end}"; read user_pub_key
+  mkdir /home/${username}/.ssh
+  {
+      echo "${user_pub_key}"
+    } > /home/${username}/.ssh/authorized_keys && \
+    chown ${username}:${username} /home/${username}/.ssh/authorized_keys
+fi
 
-# Запрос на установку приложения flask
-echo -en "\n${green}Установить приложение Flask? [Y/n]: ${end}"
-answer; if [[ $? -ne 0 ]]; then echo ""; fi
 
+echo -en "\n${green}Установить python3 и всего зависимости? [Y/n]: ${end}"
+answer
+if [[ $? -ne 0 ]]; then
 show "Установка python + dev + venv"
 sudo apt-get -y update
 sudo apt-get -y install python3 python3-venv python3-dev
+fi
+
+echo -en "\n${green}Установить python3 и всего зависимости? [Y/n]: ${end}"
+answer
+if [[ $? -ne 0 ]]; then
 show "Установка mysql server, mail agent - postfix, supervisor, nginx"
 sudo apt-get -y install mysql-server supervisor nginx
+fi
 
-echo -en "\n${cyan}Введите название приложения (en): ${end}"; read app_name
-mkdir /home/${username}/${app_name}
-cd /home/${username}/${app_name}
-mkdir /home/${username}/${app_name}/app
-{
-    echo "from flask import Flask"
-    echo "app = Flask(__name__)"
-    echo "from app import routes"
-  } > /home/${username}/${app_name}/__init__.py && \
-  chown ${username}:${username} /home/${username}/${app_name}/__init__.py
-{
-    echo "from app import app"
-    echo "@app.route('/')"
-    echo "@app.route('/index')"
-    echo "def index():"
-    echo "  return 'Hello, World!'"
-  } > /home/${username}/${app_name}/app/routes.py && \
-  chown ${username}:${username} /home/${username}/${app_name}/app/routes.py
-show "Активация виртуальной среды и установка Flask"
-python3 -m venv venv
-source venv/bin/activate
-pip3 install flask
+#echo -en "\n${cyan}Введите название приложения (en): ${end}"; read app_name
+#mkdir /home/${username}/${app_name}
+#cd /home/${username}/${app_name}
+#mkdir /home/${username}/${app_name}/app
+#{
+#    echo "from flask import Flask"
+#    echo "app = Flask(__name__)"
+#    echo "from app import routes"
+#  } > /home/${username}/${app_name}/__init__.py && \
+#  chown ${username}:${username} /home/${username}/${app_name}/__init__.py
+#{
+#    echo "from app import app"
+#    echo "@app.route('/')"
+#    echo "@app.route('/index')"
+#    echo "def index():"
+#    echo "  return 'Hello, World!'"
+#  } > /home/${username}/${app_name}/app/routes.py && \
+#  chown ${username}:${username} /home/${username}/${app_name}/app/routes.py
+#show "Активация виртуальной среды и установка Flask"
+#python3 -m venv venv
+#source venv/bin/activate
+#pip3 install flask
 
 # === Вывод данных === #
